src/log_ingestion_service/requirements.txt
fastapi
uvicorn[standard]
requests
redis
python-dotenv
pydantic
